Catch more clickbait-style headlines
Seriously, people. "Here's What You Need To Know" and "Here Are All The Things We Learned" are barely passable summaries, and are absolutely unacceptable in your titles. Why do you feel the need to keep saying your article is *here*? Maybe nobody is paying attention because your writing is atrocious.